Moved files into page directories
