+++ frontend/src/pages/MapView/Pin/index.jsx
