+++ frontend/src/pages/MapView/ResourceCard/styles.scss
