+++ frontend/src/pages/MapView/SearchBar/styles.scss
